+++ packages/stentor-context/src/ContextFactory.ts
@@ -4,11 +4,12 @@ import {
     AppRuntimeData,
     Channel,
     Context,
+    CrmService,
     Device,
     Pii,
     PIIService,
     Request,
-    SessionStore,
+    SessionStore, SMSService,
     Storage,
     UserDataRequestStatus,
     UserDataType,
@@ -22,6 +23,8 @@ import { createSessionStore } from "stentor-storage";
 export interface ContextFactoryServices {
     userStorageService: UserStorageService;
     piiService: PIIService;
+    crmService?: CrmService;
+    smsService?: SMSService;
 }
 
 export class ContextFactory {
@@ -42,7 +45,7 @@ export class ContextFactory {
         channel: Channel,
         appData?: AppRuntimeData
     ): Promise<Readonly<Context>> {
-        const { userStorageService, piiService } = services;
+        const { userStorageService, piiService, crmService, smsService } = services;
 
         const device: Device = channel.capabilities(requestBody);
 
@@ -157,7 +160,11 @@ export class ContextFactory {
             storage,
             session,
             requestUserData,
-            pii
+            pii,
+            services: {
+                crmService,
+                smsService
+            }
         };
     }