+++ packages/stentor-handler-manager/src/__test__/HandlerManager.test.ts
@@ -144,7 +144,8 @@ describe("HandlerManager", () => {
             baseContext = {
                 device,
                 storage,
-                response: new ResponseBuilder({ device })
+                response: new ResponseBuilder({ device }),
+                services: {}
             };
         });
         describe("when passed bad parameters", () => {