+++ packages/stentor-models/src/NLU/NLUService.ts
@@ -10,8 +10,22 @@ export interface NLUQueryResponse {
 }
 
 export interface NLURequestProps {
+    /**
+     * Optional locale to request, defaults to "en"
+     */
+    locale?: string;
+    /**
+     * Optional userId to pass with the request
+     */
     userId?: string;
+    /**
+     * Optional sessionId to pass with the request
+     */
     sessionId?: string;
+    /**
+     * Optional active context to pass with the request, used to weight 
+     * certain intents.
+     */
     activeContext?: ActiveContext[];
     requestAttributes?: { [key: string]: string };
 }
@@ -20,17 +34,6 @@ export interface NLURequestProps {
  * Service which can turn raw text into an intent and slots (optional).
  */
 export interface NLUService {
-    /**
-     * Send text to the NLU to return the intent and optional slots and values.
-     * 
-     * @param q 
-     * @param props - Optional properties such as userId and sessionId
-     */
     query(q: string, props?: NLURequestProps): Promise<NLUQueryResponse>;
-    /**
-     * Set the context, such as active contexts or attributes, for a user and session.  
-     * 
-     * @param props 
-     */
     setContext?(props: NLURequestProps): Promise<void>;
 }