+++ packages/stentor-runtime/src/main.ts
@@ -11,14 +11,17 @@ import { CANCEL_INTENT, STOP_INTENT } from "stentor-interaction-model";
 import {
     Channel,
     Context,
+    CrmService,
     HandlerService,
     Hooks,
     KnowledgeBaseService,
     PIIService,
     Request,
-    Response, ResponseOutput,
+    Response,
+    ResponseOutput,
     RuntimeCallback,
     RuntimeContext,
+    SMSService,
     UserStorageService
 } from "stentor-models";
 import {
@@ -72,6 +75,8 @@ export interface KnowledgeBaseDependency extends KnowledgeBaseConfig {
  */
 export interface Dependencies {
     eventService?: EventService;
+    crmService?: CrmService;
+    smsService?: SMSService;
     handlerFactory: HandlerFactory;
     handlerService: HandlerService;
     piiService?: PIIService;
@@ -113,7 +118,16 @@ export const main = async (
         throw new TypeError("Channels passed to main() was either undefined or empty.");
     }
 
-    const { eventService, userStorageService, handlerService, piiService, handlerFactory, knowledgeBaseServices } = dependencies;
+    const {
+        eventService,
+        userStorageService,
+        handlerService,
+        piiService,
+        crmService,
+        smsService,
+        handlerFactory,
+        knowledgeBaseServices
+    } = dependencies;
 
     // Step #0
     // Get the Channel
@@ -258,7 +272,9 @@ export const main = async (
             requestBody,
             {
                 userStorageService,
-                piiService
+                piiService,
+                crmService,
+                smsService
             },
             channel,
             mainContext.appData