+++ packages/stentor/CHANGELOG.md
@@ -3,6 +3,290 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.51.8](https://github.com/stentorium/stentor/compare/v1.51.7...v1.51.8) (2022-01-31)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.51.7](https://github.com/stentorium/stentor/compare/v1.51.6...v1.51.7) (2022-01-30)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.51.6](https://github.com/stentorium/stentor/compare/v1.51.5...v1.51.6) (2022-01-30)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.51.5](https://github.com/stentorium/stentor/compare/v1.51.4...v1.51.5) (2022-01-29)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.51.4](https://github.com/stentorium/stentor/compare/v1.51.3...v1.51.4) (2022-01-28)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.51.3](https://github.com/stentorium/stentor/compare/v1.51.2...v1.51.3) (2022-01-28)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.51.2](https://github.com/stentorium/stentor/compare/v1.51.1...v1.51.2) (2022-01-27)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.51.1](https://github.com/stentorium/stentor/compare/v1.51.0...v1.51.1) (2022-01-26)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+# [1.51.0](https://github.com/stentorium/stentor/compare/v1.50.27...v1.51.0) (2022-01-25)
+
+
+### Features
+
+* CRM service added to model and Assistant dependencies ([#1363](https://github.com/stentorium/stentor/issues/1363)) ([746ee27](https://github.com/stentorium/stentor/commit/746ee27005fb7d066b553d1f0c5cb1106796dd0f))
+
+
+
+
+
+## [1.50.27](https://github.com/stentorium/stentor/compare/v1.50.26...v1.50.27) (2022-01-25)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @types/aws-lambda to v8.10.92 ([#1362](https://github.com/stentorium/stentor/issues/1362)) ([ffc5fb2](https://github.com/stentorium/stentor/commit/ffc5fb2d34853e3f46e5728f431d4b30b4107675))
+
+
+
+
+
+## [1.50.26](https://github.com/stentorium/stentor/compare/v1.50.25...v1.50.26) (2022-01-24)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @types/aws-lambda to v8.10.91 ([#1355](https://github.com/stentorium/stentor/issues/1355)) ([cd2302c](https://github.com/stentorium/stentor/commit/cd2302c360dd27492edd2b611b677932075fe052))
+
+
+
+
+
+## [1.50.25](https://github.com/stentorium/stentor/compare/v1.50.24...v1.50.25) (2022-01-24)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.24](https://github.com/stentorium/stentor/compare/v1.50.23...v1.50.24) (2022-01-23)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @xapp/patterns to v1.39.67 ([d758711](https://github.com/stentorium/stentor/commit/d7587119cf568a16f3ac5d2b993c6a57f0c529be))
+
+
+
+
+
+## [1.50.23](https://github.com/stentorium/stentor/compare/v1.50.22...v1.50.23) (2022-01-23)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.22](https://github.com/stentorium/stentor/compare/v1.50.21...v1.50.22) (2022-01-22)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.21](https://github.com/stentorium/stentor/compare/v1.50.20...v1.50.21) (2022-01-21)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.20](https://github.com/stentorium/stentor/compare/v1.50.19...v1.50.20) (2022-01-20)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.19](https://github.com/stentorium/stentor/compare/v1.50.18...v1.50.19) (2022-01-20)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.18](https://github.com/stentorium/stentor/compare/v1.50.17...v1.50.18) (2022-01-19)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @types/aws-lambda to v8.10.90 ([#1350](https://github.com/stentorium/stentor/issues/1350)) ([546b390](https://github.com/stentorium/stentor/commit/546b390f809b4e76e145cfb9cf38aa7bb8281a0e))
+* **deps:** update dependency dotenv to v14 ([#1349](https://github.com/stentorium/stentor/issues/1349)) ([965c4e0](https://github.com/stentorium/stentor/commit/965c4e093729a4ea89db54bcbd5f8f42fe93c229))
+
+
+
+
+
+## [1.50.17](https://github.com/stentorium/stentor/compare/v1.50.16...v1.50.17) (2022-01-17)
+
+
+### Bug Fixes
+
+* **deps:** update dependency dotenv to v13 ([#1346](https://github.com/stentorium/stentor/issues/1346)) ([c6ccacd](https://github.com/stentorium/stentor/commit/c6ccacd7c3b652779a451baf63340180b14c435d))
+
+
+
+
+
+## [1.50.16](https://github.com/stentorium/stentor/compare/v1.50.15...v1.50.16) (2022-01-17)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.15](https://github.com/stentorium/stentor/compare/v1.50.14...v1.50.15) (2022-01-16)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.14](https://github.com/stentorium/stentor/compare/v1.50.13...v1.50.14) (2022-01-13)
+
+
+### Bug Fixes
+
+* **deps:** update dependency dotenv to v11 ([#1338](https://github.com/stentorium/stentor/issues/1338)) ([e379b90](https://github.com/stentorium/stentor/commit/e379b90bb75a4f125ae0b1cd576d091ba8f5b0c9))
+
+
+
+
+
+## [1.50.13](https://github.com/stentorium/stentor/compare/v1.50.12...v1.50.13) (2022-01-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @xapp/patterns to v1.39.50 ([9cd1df2](https://github.com/stentorium/stentor/commit/9cd1df2307a62d834286730a690da766c096e865))
+
+
+
+
+
+## [1.50.12](https://github.com/stentorium/stentor/compare/v1.50.11...v1.50.12) (2022-01-12)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.11](https://github.com/stentorium/stentor/compare/v1.50.10...v1.50.11) (2022-01-10)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.10](https://github.com/stentorium/stentor/compare/v1.50.9...v1.50.10) (2022-01-09)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @xapp/patterns to v1.39.41 ([2db0eaf](https://github.com/stentorium/stentor/commit/2db0eaf7737fa2bdb317f400954abff3313d61cd))
+
+
+
+
+
+## [1.50.9](https://github.com/stentorium/stentor/compare/v1.50.8...v1.50.9) (2022-01-07)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.8](https://github.com/stentorium/stentor/compare/v1.50.7...v1.50.8) (2022-01-06)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.7](https://github.com/stentorium/stentor/compare/v1.50.6...v1.50.7) (2022-01-04)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.6](https://github.com/stentorium/stentor/compare/v1.50.5...v1.50.6) (2022-01-03)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
+## [1.50.5](https://github.com/stentorium/stentor/compare/v1.50.4...v1.50.5) (2021-12-30)
+
+**Note:** Version bump only for package stentor
+
+
+
+
+
 ## [1.50.4](https://github.com/stentorium/stentor/compare/v1.50.3...v1.50.4) (2021-12-29)
 
 **Note:** Version bump only for package stentor